Drop support for ownCloud <10, Nextcloud <13, and PHP 5.6 (and IE9)
These cloud and PHP versions have reached the end-of-life years ago, so
if people are still using them, then they really should move on. In
practice, this also means that we no longer need to support Internet
Explorer 9: It does not work on NC13 at all, and on oC10 it "works" but
the layout is badly broken also in all the built-in apps.