Remove the obsolete IE SVG hack
The core function replaceSVG used in the hack has had an empty
implementation already since ownCloud 9.0. There are still some
problems with the play icon size on IE10 and IE11, but these obviously
can't be solved by calling an empty function.

Oddly, it seems to be so, that if the Music app is (re)loaded with an
address pointing to some track/album/artist, then the play icons are
always incorrectly sized on IE. The page needs to be reloaded while the
address does not point to anything playable to fix the issue.