Remove the css opacity fallback for old IE versions
For every opacity property in the css files, there were two distinct
fallback properties needed to support old IE versions. But the thing
is, we have never or at least for many years supported IE versions
which would need those fallbacks: Already IE9 should support the
standard opacity property. One of the fallbacks was intended for IE8
and the other for IE versions 5-7.

As we now support only IE10+, it is more than a good time to get rid of
these.