Ampache: Fix handling of `offset` argument in the action `playlists`
The action `playlists` appends the the pseudo-playlist "All tracks" after the
actual playlists. The logic to add this list didn't work correctly when the
`offset` argument had been given: depending on the given `limit`, either the
"All tracks" could be left out when it would have fit the result set, or then
it was appended even when `offset` was so high that the result set should
have been empty.

refs #768