Ampache: Provide limited support for actions `download` and `stream`
The Ampache API version 400001 has added, among others, the new actions
`download` and `stream`. Although we are mostly targeting the API version
350001, support for these two was added. The `download` action is now used
instead of our former proprietary action `_play` to construct the URL
included in all returned `song` entities.

To be fully compliant with the API specification on these actions, both of
them should be able to transcode the content to desidered file format. In
addition, the `stream` action should be able to do the transcoding with a set
maximum bit rate, and should be able to stream the file from a given time
offset. Our implementation doesn't support any of these.

refs #768