Add option to scan only the specified folder with `occ music:scan`
By supplying an optional command line argument --folder=<target_path>, the
scanning from the command line can limited to the specified folder and its
subfolders. The path must be given relative to the user's home folder. If the
given path does not exist, or is not within the users configured music
library path, then nothing will be found.

Using this option probably makes most sense together with the option
--rescan, but it does work regardless.

refs #771