Use format "h:mm:ss" for play times longer than one hour
Previously, all play times were always formatted like "m:ss". This
wasn't very user friendly for playlist total time, where the time may
be several hours or more. Now the formatting includes also the hours
part if the formatted time is more than one hour. For time periods
shorter than one hour, the formatting is like before.

Also, removed the integer conversion with the ceiling function in the
back-end class DetailsHelper. This is basically part of the formatting,
which is the responsibility of the front-end.