Initial support for importing playlist contents from a m3u file
The file references in the file must be given using relative paths. If
some of the listed files cannot be found from the user's music library,
then those files are skipped.

The file must have UTF8 encoding (if not plain ASCII). Probably more
correct logic would be to assume Latin-1 if no other encoding is
defined in the file.