Include EXTM3U fields in the exported playlist files
It would be nice to include also the artist name on the #EXTINF fields.
However, that would currently require digging the artist entities
separately from the ArtistBusinessLayer. Maybe it would be better to
refactor the backend so that artist names would be glued to every Track
object by the TrackMapper.