Tweak the "import from file" icon
Both "export to file" and "import from file" now have the file symbol
on the right and arrow on the left. It's only the direction of the
arrow that is different between these two. Previously, the arrow pointed
from left to right on both icons, and the file symbol was in the
different ends of the arrow in the two icons.

Having the arrow point to reverse direction makes it more clear that
these two operations are the reverse operations of each other.