Add REST API endpoint to parse any playlist file
The REST API route `/api/playlists/file/{fileId}` can be used to parse
a .m3u or .m3u8 file, without doing anything else. Unlike on import
from a playlist file, the contained files do not have to reside within
the music library of the user. It is still checked that the files can
be found from the user's cloud storage.

This shall come into use when we add the support to open playlist files
directly from the Files app.