Files music player: Cache the latest playlist file contents fro reuse
Playlist file reading now happens via the module
OCA.Music.PlaylistFileService both when playing the list and when
viewing its contents in the sidebar.

The module caches the latest loaded playlist contents. Hence, if the
user starts playing a list, and then clicks the "Show list" button, the
list does not need to be reloaded but can be shown immediately.
Similarly, if the user first views the contents of a playlist and then
decides to play it, the playing can start immediately.