Files music player: Make the Playlist details tab interactive
It is now possible to jump to any track of a playlist by clicking it in
the Playlist tab of the details sidebar. It is even possible to start
playing a list which is not already playing.

To make this work, the main.js of the embedded player had to be
refactored so that the context reference is not used. All the needed
data could be obtained from the global object OCA.Files.App.fileList,
instead.