Fix embedded player being loaded on the login page
There was such a problem that our embedded files music player could get
loaded on the login page, in case the URL contained a query part with
redirect to the files app. In that case, our regex incorrectly
concluded that the files app is being loaded, and embedded player is
needed.

Login URL with a redirect arg is used when the user session has
expired, and the user is moved to the login page to reauthenticate.