Files music player: Scroll to playlist when "Show playlist" clicked
There was a bug which caused an empty/corrupted sidebar to be shown.
This happened when the "Show playlist" button was pressed for such
playlist which was not loaded by the Files app. This could be because
of lazy loading or because the playing list was not in the currently
viewed folder. The following changes were made to fix these two cases:

a) If the currently viewed folder is not the parent folder of the
playlist, the "Show playlist" does nothing but shows an error note.

b) The target playlist file is scrolled to the viewport before opening
the details pane, thus ensuring that it is already loaded by Files.