Refactor the creating of the parsed playlist track download URL
The result of the API endpoint `/api/playlists/file/{fileId}` was
changed slightly to make it more similar with the model of the Files
app. This allows us to create the file download URL the same way for the
playlist tracks as we are using on the tracks played individually in
the Files app.