Files music player: Formatting on the file names in the playlist tab
The track names shown in the playlist tab view are now obtained from
the file names the same way as elsewhere: the file extension is
dropped, as well as any prefix which looks like a track number.

Also, the track's number on the list is now shown with a gray font.
This is implemented using the experimental ::marker CSS pseudo element.
Currently, this works on FIrefox. On Chrome, it works only if the
experimental web features are enabled from the chrome://flags.