Show the "more" icon for playlist in navpane only on hover or focus
The "more" icon on the playlist items on the navigation pane is now
hidden by default. The icon becomes visible when the playlist item is
hovered or the playlist in question is active.

On the other hand, the icon now has a bit more contrast (higher
opacity) to make it more visible.