Files music player: Use track title from #EXTINF field if available
The #EXTINF fields of the m3u(8) files are now parsed along with the
actual playlist entries. If the field is present and contains a track
caption, then this is used as the title of the playlist item on the
playlist tab view. Otherwise, the title is generated from the file name
as before.