Enable importing playlist file from the Files app
In the Files music player, there is now a popup menu in the place where
there used to be the "Show playlist" button. In this menu, there are
two options:
- Show playlist
- Import to Music

The second option creates a new playlist with the name of the playlist
file, and imports the file's content to this list.