Files music player: Disable playlist menu items if action not possible
Showing the playlist contents in the sidebar is not possible when
viewing folder other than the parent folder of the playlist. This is an
architectural limitation in the sidebar of the Files app. This is now
indicated by disabling the related action in the popup menu.

Also, the action "Import to Music" makes no sense if the tracks of the
playlist do not belong to the user music library. In such case, this
menu option is now disabled.

In case the menu item is disabled, a tooltip tells the reason for
disabling. Furthermore, in case only some of the playlist tracks can be
imported, also this is indicated by a tooltip.