Files music player: Fix player being broken on link shared folders
The player had got broken on publicly shared folders with the commit
552dfc0f "Files music player: Make the Playlist details tab
interactive".

The problem was that the OCA.Files.App is not available while viewing a
public share. Hence, we need to obtain the fileList reference from the
passed in context after all. It is now obtained from OCA.Files.App only
in the case when the playback is started from the sidebar playlist view.
The sidebar is available only within the Files app, so this is safe.

Playing playlist files within the publicly shared folder still doesn't
work. This is because the playlist parsing happens using the API
endpoint `/music/api/playlists/{fileId}/parse` which is not public.