Fix wrong artist image sometimes being shown
As it may take some time to load the artist image, it was possible to
end up in a situation, where the artist details pane show name of one
artist but image of another. This could happen when the artist shown in
the details pane was changed while the image of the previous artist was
still being fetched.