Remember the play offset when navigating from Files player to Music
When the user clicks the track info area in the Files music player, the
playback continues within the Music app. Previously, the playback in Music
started from the beginning of the clicked track, but now also the play
position is transferred, and the playback continues from the same position
where it was left on Files.