Enable navigation from track details to album and artist details
The track details now has th "i" icon after the album name as well as the
artist and album artist names. Clicking the detail line with the icon will
show the respective album or artist details pane.

What is noteworthy, is that this allows viewing the artist details of also
those performing artists who don't have a single album in the collection.