Fix rendering of lyrics and other track details with "CR only" new lines
The "CR only" new line style shouldn't be used on any modern systems,
but apparently, files got from iTunes may use this style for the
embedded lyrics. Meanwhile, the web standards do not support this
style. Hence, lyrics with "CR only" new lines were rendered on web
browsers as if they had no line breaks at all.

Now all the new lines of the track details are normalized to the "LF
only" style already on the back-end. This is applied also to the lyrics
provided via the Subsonic API, as it was tested that at least Subsonic
Music Streamer and DSub can't handle the "CR only" style, either.

refs #780