Fix navigating from Files player to Music on new NC versions after 447a2
After the commit 447a2f0e, navigating from the Files music player to
the same track on the Music app did not work at least on Nextcloud 17
and newer. On Nextcloud 13, at least, it worked fine.

The problem was, that the URL query part '?offset=xxxxx' was included
in the REST API call used to fetch the track matching the given file
ID. This broke the API call, as RESTangular URL-encoded the '?', and
query part no longer looked like a query part. Presumably, this worked
somehow accidentally on the older NC versions. It may be that they did
some automatic URL-decoding before parsing the query part, which would
be basically wrong.