Subsonic: Show also featured albums when viewing an artist
Even though our implementation of the method getArtists returns only
artists who have albums, there are still many ways for the user to view
those artists who are featured on some album but do not have any albums
of their own. On DSub, at least these are possible:
- use search to find the artist
- use "Similar Artists" feature
- select "Show Artist" for a track from the current play queue

Previously, as we were not returning any albums for the artists with no
*own* albums, DSub failed to show the artist at all. That is, there
wasn't event the info from Last.fm shown in that case.