Subsonic: Empty response from getArtistInfo if called for folder
Our logic for getting the artist info from LastFM is built on the
artists IDs, so it doesn't work when browsing by folders (even if the
folder name would match some artist name). For now, just return an
empty response if getArtistInfo is called with a folder ID as an
argument.

Previously, calling the method with folder ID (usually) caused an
exception which resulted in to a Subsonic error response. This made
DSub hang when it tried to show the artist info for the folder.