Ampache: Support for the action 'stats'
Not all of the possible filters of the action are supported at this
time, because the needed data is not available in our database. The
currently supported filters are 'newest', 'flagged', and 'random'.

The helper function `injectAlbum` was removed, because the album of
each track can be fetched within `renderSongs`. There was only a single
case where "commonAlbum" was being used with `injectAlbum`, and this
occasion has now been replaced with a simple inlined loop.