Subsonic: Do not attempt to convert any timestamps to local timezone
It is not stated in the Subsonic API documentation, but apparently also
the real Subsonic server uses UTC time on all timestamps. What
previously made me believe otherwise, was just the DSub issue
https://github.com/daneren2005/Subsonic/issues/1001. Even DSub expects
to get UTC timestamps on the bookmarks.

Also, the timestamp formatting was changed to the "zulu format" with
milliseconds included, because this is also what the real Subsonic
server uses (at least on http://demo.subsonic.org).