Ampache: Show dummy song if playlist contains an invalid track ID
Previously, having an invalid track ID on playlist was treated by just
skipping that track when listing the playlist songs. This was
confusing, because it made the track ordinals in `playlist_edit` and
`playlist_remove_song` to differ from the ones expected by the user
based on the listed songs.

Now we instead show an empty dummy song on the position where the
invalid ID is.