Replace use of the legacy core functions within the front-end templates
The front-end templates used to rely on the global php functions `p` and
`print_unescaped` which are defined in the core file
`lib\private\legacy\template\functions.php`. The fact that these are
found under "private" and "legacy" strongly suggest that applications
should migrate away from using these. Furthermore, Scrutinizer couldn't
find the definitions for these functions as they are not under the
public API.

As a replacement, we now define own utilities for the use of the
templates in \OCA\Music\Utility\HtmlUtil. Our own utilities mostly
operate on a bit higher level than the replaced core functions,
reducing repetition within the templates a bit.