Allow removing files in excluded paths with occ music:scan --remove-obsolete
The option `--remove-obsolete` on the command `occ music:scan` now removes
also any tracks which now reside in an excluded folder. This is in addition
to its former function of removing such tracks which are no longer accessible
by the user at all.

The new logic for the `--remove-obsolete` switch is also a lot more efficient
on huge libraries. As an example, checking for the obsolete files now took
~40 seconds on library of more than 60k tracks. With the old implementation,
I got tired of waiting and terminated the test after 10 minutes, so it was at
least 15 times slower and maybe much more.

refs #762