Fix exclude path with wild cards being treated as if it had trailing *
Any exclude path pattern containing wild card characters was incorrectly
matched to also such paths, which began with the given pattern, but then had
some more characters after that. That is, the pattern was treated as if it had
a trailing '*' wild card even when it didn't. For example, pattern
/*/abc
matched also the file
/asdf/abcdef/some_file.mp3

refs #762