Fix javascript breaking on minification
The explicit annotation of the injected dependencies was missing from
the controller of the InViewObserver directive.

Also, we now added the `ng-strict-di` flag for the ng-app element. This
breaks the app always if the explicit annotation of any dependencies is
missing, even when the minification is not being used. This should make
it much easier to spot errors like the one fixed here, already before
submitting the faulty code.