Enable using webpack to bundle the app .js and .css
With this configuration, it is possible to create working application
bundle .js and .css files by running `npm run build` under the build
directory. The bundle includes both the application sources and the js
vendor libraries. The main.php has been modified to use the bundles
created by webpack. The files-music-player has not been touched, yet.

One noteworthy thing is that we now bundle along the lodash.js library
which is used instead of underscore.js shipped with the core. This is
because the Restangular library has dependency to lodash.