Couple of js changes needed to migrate from underscore.js to lodash.js
The "object oriented syntax" like `_(list).function()` works a bit
differently on lodash.js as compared to underscore.js: In lodash, it
implicitly also starts chaining, and to get the final value, another
function call `.value()` would be needed (unless the last function is
one returning an individual list value). We now replaced the
`_(list).function()` syntax with the traditional `_.function(list)`
syntax in those places where this made a difference.

Also, the third argument of the `indexOf` function has different
meaning on lodash compared to underscore: In lodash, it is a start
index while in underscore it was a boolean telling if the list is
sorted. The lodash library provides another function `sortedIndexOf`
for the searching in a sorted list.