Drop bower and use npm as the only package management tool
Everything previously downloaded with bower to folder js/vendor is now
downloaded with npm to folder build/node_modules. The libraries under
js/vendor used to be also version controlled within this repository,
but for the node_modules, we trust that npm can always get the correct
version.

The folder js/vendor now contains only those vendor libraries which are
not available via a package management tool.

While making this change, we also updated the Angular version from 1.8.0
to 1.8.1.