Prevent lodash from leaking to the global window scope
This is a well-known issue when using lodash+webpack, and has been
reported e.g. in https://github.com/lodash/lodash/issues/3052. The
work-around implemented here is the one suggested on the linked issue.

Leaking to global scope was a problem at least with the embedded
files-music-player, as the Files app assumed to find underscore.js (with
e.g. `pluck` function) from the global name `_`.