Include own version of placeholder.js to our webpack
As we are using our own instance of jQuery, it now does not have the
placeholder plugin implemented by the cloud core. Hence, the placeholders
were not shown on the albums without a cover image, and an error was printed
to the browser console, instead.

To overcome this, we now install our own instance of the placeholder.js. This
has been ripped from Nextcloud 20.