Fix yet another problem from the underscore.js -> lodash.js migration
In lodash, the functions _.first and _.tail don't have an overload which would
take a second parameter. These functionalities are achieved with the
functions _.take and _.drop, respectively. (The same would be true for the
function _.last, but we did not use the overload of this function anywhere.)

The small change made on the artistdetailscontroller.js is not related to
the problem described above. This is just a small rationalization, leading to
neater code and possibly better performance.