Don't let aurora or js-cookie libs leak to the global namespace
Utilize webpack's ProvidePlugin also on the vendor librarires `aurora` and
`js-cookie` so that we don't have to leak the `AV` and `Cookies` objects to
the global `window` namespace.

Now the only object still leaked to the global scope is `angular`. In
addition, we install our own modules to the namespace OCA.Music, which
wouldn't be strictly necessary, but probably will make no harm, as this
namespace is kind of reserved for us.