Fix a navigation bar layout issue visible on NC20 and OC10.4
Nextcloud 20 and ownCloud 10.4 highlight the active view in the navigation
pane with a background color. This made it visible that the higlighted (and
clickable) area of the list item did not fill the whole row, but there was a
44x margin on the right side. This was left to fit the "actions" button which
opens the popup menu.

Making the clickable and highlightable area to span the whole navigation pane
width required quite a bit of changes on the structure of the navigation pane
items. Most notably, the first element within the <li> element can not be an
<a> element, because making the popup menu work correctly inside an <a>
element seemed to be impossible. This, on the other hand, forced us to define
many css rules which previously came from the cloud core.