Keep `css` and `js` folders in the release package to fix a showstopper on ownCloud
For some reason, ownCloud (all supported versions) requires that the
app must have `css` and `js` folders in its root, even if they are
empty. If the folders are not there, the .css and .js files will not be
loaded from the `dist` folder, and the web UI cannot be used. Nextcloud
has no such issue.

The .gitattributes was now changed so that only the contents of those
folders are excluded from the release, but the folders themselves are
included.

refs #792