Attempt to prevent extensive logging in case of unexpected DB content
According to reports, the modified lines could produce a huge amount of
log output in case the albumId in question was not found from the
array. On some systems, the whole array was then printed to the log,
and the array may contain thousands of items.

The situation described above shouldn't normally happen because it
would require the database to contain an album which has no tracks at
all. Such albums should not exist, and the background cleanup task
should take care of removing them if they still somehow came to be.
Also, I couldn't reproduce the extensive logging issue by creating a
bogus album with not tracks directly to the DB; it may depend on the
version and configuration of the cloud or PHP.

refs #793