Cache busting for the webpack assets
The .js and .css bundles created by webpack now have a content hash
attached to their name. This should ensure that they get reloaded after
any change. Previously, especially Chrome was reluctant to
reload changed bundles without the forced reload function. Firefox had
the same issue with the .css but not with the .js. The problem was
especially bad on the mobile browsers which apparently have no easy way
to trigger the forced reload.

As the bundles now change name on each change, the CleanWebpackPlugin
was added to clean any obsolete files from the output folder. The
WebpackAssetsManifest plugin was added to create manifest.json on
build, to enable mapping from the webpack entry names to the
corresponding file names including the hash. This is done in the
functions HtmlUtil::addWebpackScript and HtmlUtil::addWebpackStyle,
which are now used in place of \OCP\Util::addScript and
\OCP\Util::addStyle.