Fix `occ` commands failing when the Nextcloud Mail is installed
The Nextcloud Mail (at least the latest version on NC20) ships its own
version of the Symfony library. When the app is installed, that version will
be used also for the `occ` commands of the Music app. This version is more
strict on the return value of the functions overriding Command::execute(),
throwing an error if the function does not return an integer value.

Fixed by adding an integer return value to each override of the
Command::execute() function. Fortunately, this seems to be allowed also by
the Symfony library versions shipped with older versions of Nextcloud and
ownCloud.

refs #799