Do not build the webpack bundles as part of `make release`
This eases the pain of making the release for me. The problem is that I don't
currently have an environment where both the webpack build and creating the
zip file would work correctly: Webpack building succeeds in Windows
PowerShell, but fails on my Linux system (probably should update the Linux to
newer version to get up-to-date tools there), and results in slightly
misbehaving package when run from Cygwin. But the Linux-like command line
utility `zip`, on the other hand, is available on Cygwin and Linux but not on
PowerShell.