Use only the cache-friendly image file paths
Previously, the image files referenced within the css files were substituted
with their hashed alternatives under dist/img by the webpack compilation. The
images referenced from the PHP templates or from the js code, on the other
hand, were the original "source" versions of the svg files. Hence, we had to
ship each svg image twice with the application.

Now the images with hashed names are used also within the js and php code.
Hence, the original source versions of the images can be stripped off the
delivery package.