Update jQuery to version 3.5.1
Previously, we were forced to use an older version since using the
jQuery version 3.5.x caused layout of the playlist view to break down.
Now this was found to be caused by malformed HTML in the playlist view:
the HTML specification does not allow the <div> tag to be written as
self-closing. After fixing the HTML, the latest jQuery version 3.5.1
can be used without any issues.