Remove Util::arrayGetOrDefault and use the ?? operator instead
The null coalescing operator ?? was added in PHP 7.0 and it does
basically exactly what we were doing with our own utility function
Util::arrayGetOrDefault (although ?? is more versatile, and can be used
with objects, too). The ?? operator even supports the case of nested
arrays with potentially missing middle nodes; support for this case was
added to Util::arrayGetOrDefault just recently.

As we dropped support for PHP 5.x on v0.15.0 a couple of months ago, we
are now free to utilize any PHP7-specific features. Using the ??
operator saves us some code and may potentially perform better than our
custom function.