Move own PHP sources to the subdirectory `lib`
This is the more modern alternative to organize the application project
in ownCloud or Nextcloud, and leads to much cleaner directory
structure. When organizing the project files like this, the clouds
require also that the source files and folders follow the PSR-4 naming
conventions. That is, the directories matching the namespaces and files
matching the classes must now have name with identical capitalization
as compared to the namespace/class names. This is in contrast with the
previous practice of using all-lowercase file names.

Two classes were renamed during the process to follow the "best
practices":
APISerializer -> ApiSerializer
XMLResponse -> XmlResponse