Use strict typing on all PHP sources
Quite a few tweaks had to be made on the code to function properly with
the strict typing. Also, a bunch of new type declarations were added;
some but not all of these were necessary to make the logic work with
the strict typing.

There is also a good chance that some problems caused by the strict
typing were still left hiding. Unfortunately, such problems are found on
an interpreted language like PHP only upon running the problematic line
of code. Hence, we now need to take a leap of faith, and hope that we
can find any issues before making the next release.