One more CSP rule needed for HLS streaming on Chrome
The "media-src: 'blob:'" seems to be necessary on Chrome but not on
Firefox. Also the "media-src: 'data:'" is needed on Chrome for HLS
streaming; it used to be necesary even for local playback but that
doesn't seem to be the case any more. This has probably changed when we
dropped the SoundManager2 library.

refs #784