Enable jumping to the beginning of the track also with Aurora
The Aurora.js library used as a fallback playing method does not support
seeking (*). However, we can simulate the seeking to the 0-offset by
stopping and reinitializing the player state. This workaround is now
used when seeking to 0 is requested while playing a track with Aurora.
Hence, it is now always possible to jump to the beginning of the current
track by pressing the "Skip previous" button.

(*) Actually the problem is in the plug-ins of the library: The FLAC
plugin does not implement seeking and the MP3 plugin has an unreliable
implementation for it.