Add "repeat one" as a third state for the "Repeat" button
In the "repeat one" mode, the currently playing track is played again
from the beginning each time it finishes playing. However, using the
"skip next" button still jumps to the next track, and repeated song is
only once in the play history, regardless of how many times it got
played.

refs #808