Workaround to show the cover images on the mediaSession also on Firefox
Unlike Chrome and Edge, the Firefox browser loads the cover image given
to the mediaSession API in a distinct context where the normal browser
session cookies are not present. Because of this, the album covers
couldn't be shown by the OS player controls with Firefox, as our album
cover URLs required the user to be logged in.

This has now been worked around by making the REST endpoints serving
the cover images technically public. However, if there is no logged in
user, then we require a temporary token as an URL argument. This token
is given as an additional property within the response to the
`prepare_collection` request. Only the latest obtained token per user
is valid.

The system is not perfect, as it will break down if the user uses
simultaneously multiple browsers. However, this shouldn't be too
serious an issue as it will only cause the cover image loading to fail
within the mediaSession context (like previously happened every time).