Do not use redirection when the Firefox mediaSession requests album art
The redirection to the cacheable route did not work because the coverToken
argument was not included in the redirection URL. This could have been fixed
by relaying the token also to the redirection URL, but it is unclear if using
the cacheable URLs with the Firefox mediaSession would have any benefit. And
even at best, the difference would be minuscule.