Use the own API endpoint instead of WebDAV to play files on the web UI
The WebDAV was first taken inot use here six years ago because it provided
support for the range requests. However, these have been supported also in
our own FileResponse responses already for many years. Hence, there is no
compelling reason to use the WebDAV, and using our own REST API helps us to
simplify the code. Also, the playback can now be started ~100-300 ms faster
after the user action as the additional REST API call to get the file path is
no longer needed (although noticing such a small difference is not easy).

Also, the REST endpoint `/file/{fileId}/download` was streamlined a bit. It
was fully functional (but unused) already earlier, but now it has been
refactored to a simpler form. Previously, the function checked that the given
file ID actually belonged to an indexed music track. However, there should be
no real security reasons to do this: when obtaining the file via the
UserFolder class, only the files accessible by the logged in user are
available. And these can anyway be accessed through the Files app and WebDAV
APIs (if just the path is known).

Furthermore, the debouncing related to changing of the playing track on the
front-end is now implemented using the lodash function _.debounce instead of
providing an own hand-crafted solution.