Fix small issues with seeking on the web UI
After the commit 797a01ee9, there has been such an issue, that the cursor
icon on top of the seek bar was shown as an "arrow" instead of "pointer" while
playing the first track during a session on the Music app web UI, even if the
seeking was supported. This has now been fixed.

Also, seeking is now disabled and the mouse cursor shown as "arrow" while the
track is not yet ready to play. Previously, seeking was possible but could
lead the playback to jump into an unexpected position.