Fix the support for IE10 and IE11 broken since v0.17.0
The commit 44e7ba9c first released in v0.17.0 broke the support for any
IE version, because it brought in a vendor source file which used the
keyword "const" introduced in ECMAScript 2015. We now use babel with
webpack to transpile the code to more compatible format. To enable
this, we also had to slightly modify the vendor library aurora/flac.js:
The babel adds "use strict" to all js files, and the library had a
single line of code not complying with the strict rules. Monkey
patching this script shouldn't cause any issues, as the library is
abandoned by its creators anyway.

Furthermore, our webpack entry point code broke the support for IE10
because it tried to load the jquery-initialize library which failed on
browsers not providing the MutationObserver class. In places using the
jquery-initialize it was already taken into account, that the library
is not available on such browsers.