First implementation of stricter content security policy
Our previous (temporary) implementation for allowing the streaming from
external sources despite of CSP was to to white-list every possible HTTP and
HTTPS URL. This was a potential security risk (even though CSP is just the
second line of defense) and never intended to be the final solution.

Now we maintain separately a list of allowed external domains. In this first
version, the list is updated automatically when importing radio channels, but
there is not yet any manual means to alter it.