Fix seconds part of play time showing incorrectly in range [59.5, 60.0)
This effectively rolls back the commit 989d5fb3. Using `round` instead
of `floor` for the seconds of the playtime wasn't such a great idea
after all, because it made the time stamp to show up like "0:60" for a
half a second before it changed to "1:00". It would have been possible
to keep the rounding to closest by altering the logic for minutes and
seconds parts, but the rounding had also a second minor issue: the play
time changed from 0:00 to 0:01 already after playing for half a second
which isn't quite how a time display is assumed to work, normally.

refs #814