Fix the error notification shown if playing a URL fails
The code assumed that the file path could be found as an URL argument
in the URL that failed to play. However, this no longer holds true even
for local files, let alone the external stream URLs which may have what
ever format. We now show the whole URL on file play error.