Fix a layout issue with the actions menu on the navigation pane
The layout got messed up when the actions menu was opened for a
playlist which was not the active view, and then another playlist below
the one with open menu was hovered over. In this case, the open popup
menu got dimmed, and the hovered playlist item in the navigation pane
got drawn over the popup menu.

The issue fixed here is not directly related to the Internet radio
feature, but in addition to the playlists, it affects also the popup
menu of the Internet radio navigation item. Hence, it is easier to fix
it within the feature branch.