Allow all non-HLS radio sources by default, HLS sources enabled via config.php
It turned out to be a bad idea to allow ordinary users to select the
allowed radio streaming sources. This is because it was easy to add a
great number of white-listed source addresses and eventually exceed the
HTTP header size limit. If this happened, the Music app couldn't be
opened any longer, but only an unhelpful error message was shown. To
make the problem difficult to avoid, we don't know the exact limit as
it may depend on many things on each server setup like HTTP server, its
settings, and any proxies.

So now, we let only the system admin select the allowed sources by
editing the config.php file. By default, we allow all non-HLS streaming
sources by adding `http://*:*` and `https://*:*` to the `media-src`
property of the CSP. On the other hand, by default, we allow no HLS
sources, as this requires allowing the source as `connect-src` which
might be potentially more dangerous.

The hls.js library has been configured to not use WebWorker. This way,
we don't need to allow the worker-src blob: in the CSP (or alternatively,
child-src or script-src). According the documentation, using the WebWorker
improves performance, but there doesn't seem to be much issues with the
cpu load in our use case even when the worker is disalbed (tested on
Firefox and Chrome).