Add `created` and `updated` timestamps for all library entities
At this stage, the timestamps are just stored to the database but not
yet used for anything (apart from the previously existing use cases with
the playlists and bookmarks).

Adding and updating the timestamps is handled automatically by the
BaseMapper, the derived Mapper and BusinessLayer classes don't need to
worry about these. All the Entity types using a mapper derived from
BaseMapper just need to have the fields `created` and `updated`.