Ampache: Add support for the `add` and `update` filters
Filtering according the addition and update timestamps is possible with
the actions `get_indexes`, `artists`, `albums`, `songs`, and
`playlists`.

The timestamp may be given in any format understood by the PHP DateTime
constructor. For example ´2021-01-31T17:56:00.123', '2021-01-31 17:56',
and '2021-01-31' are all accepted formats. When the time of the day is
not given, the 00:00:00 is implicitly assumed.

Giving a single timestamp is interpreted as the minimum allowed add or
update time, but it is also possible to give the accepted range using
the syntax 'min_time/max_time', e.g. '2021-01-28/2021-01-31'.